+++ graphs/topological_sort/__init__.py
@@ -1,6 +1,6 @@
 from basic_data_structures.linked_list import LinkedList, Node, list_insert
-from graphs.dfs import Counter, WHITE, GRAY, BLACK
 from graphs import Graph, Vertex
+from graphs.search import Counter, WHITE, GRAY, BLACK
 
 
 def topological_sort(G):
